fdfd

+++ src/js/boardFunctions.js
@@ -45,7 +45,7 @@ export const updateChildren = async (currentCard, relativeCards, token) => {
 					let childMoment = moment(currentCard.due_date).utc().add(childCard.difference, 'M')
 					const decimals = childCard.difference.toString().split('.')[1]
 					console.log(decimals)
-					const dec = childCard.difference >= 0 ? parseInt(`0.${decimals}`) : -parseInt(`0.${decimals}`)
+					const dec = childCard.difference >= 0 ? parseFloat(`0.${decimals}`) : -parseFloat(`0.${decimals}`)
 					console.log(dec)
 					const daysRest = Math.floor(dec * 30)
 					console.log(daysRest)
